fix: add test container support for KeycloakInstanceWrapper
