+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-identity-keycloak/src/test/java/org/activiti/cloud/services/identity/keycloak/KeycloakInstanceWrapperAdminClientIT.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2017-2020 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.activiti.cloud.services.identity.keycloak;
+
+import org.springframework.test.context.TestPropertySource;
+
+@TestPropertySource("classpath:application-admin-client.properties")
+public class KeycloakInstanceWrapperAdminClientIT extends KeycloakInstanceWrapperTestSupport {
+
+}
