+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-identity-keycloak/src/test/java/org/activiti/cloud/services/identity/keycloak/KeycloakInstanceWrapperClientCredentialsIT.java
@@ -15,23 +15,9 @@
  */
 package org.activiti.cloud.services.identity.keycloak;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 
-import static org.assertj.core.api.Assertions.*;
+@TestPropertySource("classpath:application-client-credentials.properties")
+public class KeycloakInstanceWrapperClientCredentialsIT extends KeycloakInstanceWrapperTestSupport {
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
-@TestPropertySource("classpath:application.properties")
-public class KeycloakInstanceWrapperIT {
-
-    @Autowired
-    private KeycloakInstanceWrapper keycloakInstanceWrapper;
-
-    @Test
-    public void shouldWireWrapper() {
-        assertThat(keycloakInstanceWrapper).isNotNull();
-        assertThat(keycloakInstanceWrapper.getRealm()).isNotNull();
-    }
 }
