+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-identity-keycloak/src/test/java/org/activiti/cloud/services/identity/keycloak/KeycloakInstanceWrapperTestSupport.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2017-2020 Alfresco Software, Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.activiti.cloud.services.identity.keycloak;
+
+import org.activiti.cloud.services.test.containers.KeycloakContainerApplicationInitializer;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
+@ContextConfiguration(initializers = {KeycloakContainerApplicationInitializer.class})
+public abstract class KeycloakInstanceWrapperTestSupport {
+
+    @Autowired
+    private KeycloakInstanceWrapper keycloakInstanceWrapper;
+
+    @Test
+    public void shouldWireWrapper() {
+        assertThat(keycloakInstanceWrapper).isNotNull();
+        assertThat(keycloakInstanceWrapper.getRealm()).isNotNull();
+        assertThat(keycloakInstanceWrapper.getRealm().groups().groups()).hasSize(2);
+        assertThat(keycloakInstanceWrapper.getRealm().users().list()).hasSize(8);
+    }
+}
