+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-identity-keycloak/src/test/resources/application-admin-client.properties
@@ -4,3 +4,4 @@ keycloak.realm=activiti
 activiti.keycloak.admin-client-app=admin-cli
 activiti.keycloak.client-user=client
 activiti.keycloak.client-password=client
+activiti.keycloak.grant-type=password
