+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-identity-keycloak/src/test/resources/application-client-credentials.properties
@@ -1,0 +1,6 @@
+keycloak.auth-server-url=http://localhost:8180/auth
+keycloak.realm=activiti
+
+activiti.keycloak.client-id=activiti-keycloak
+activiti.keycloak.client-secret=545bc187-f10f-41f9-8d5f-cfca3dbada9c
+activiti.keycloak.grant-type=client_credentials
